Move `CompilingPass` input type to a GAT/TAT
Following a discussion had with the core maintainer of the project, we
wondered if it would be possible (and also preferable, but that is
another question), to move the input type of a `CompilingPass` to be a
generic type associated to the trait implementation rather than a pure
generics of the trait.

The obvious advantages of having generics are that it is more intuitive
for developers. It is also less of a concrete type, so no lifetimes are
necessary. The downsides are that we never really need the same type
implementing `CompilingPass` to also implement the same pass on another
type that cannot be trivially converted using a method such as
`.into()`.

On the other hand, the upside of using a type associated to the trait is
that it is more coherent with the fact that the output type is already
specified that way. It also formally restricts the fact that a single
structure will only ever perform one type of pass over one type of input
data. The downside is that in cases where it is necessary to specify a
lifetime (notably when references, such as `&str` are involved), we need
to move the TAT to be a Generic Associated Type (GAT), with an
associated generic lifetime. The syntax is a bit less common, but not
much more verbose. The generic lifetime still needs to be specified even
for types that do not need it however.

This commit is a proof of work of this modification, to invite
discussion among the developers, and show ways that the
`CompilingPass` could be thought of.

Signed-off-by: Amélie "Lymkwi" González <lymkwi@vulpinecitrus.info>